requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
click==8.1.3
colorama==0.4.6
fastapi==0.96.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
psycopg2==2.9.6
pycodestyle==2.10.0
pydantic==1.10.9
pyflakes==3.0.1
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
